requirements.txt
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==0.9.1
urllib3==1.26.5
